+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 
 ### What's Changed
 * Added: Calculation of history values not provided by the battery by @mr-manuel
+* Added: Error handling for `config.ini` by @mr-manuel
 * Added: Felicity BMS by @versager
 * Added: JKBMS CAN - Extended protocol with version V2 by @Hooorny and @mr-manuel
 * Added: Kilovault HLX+ BMS by @alexphredorg